+++ robotcode/language_server/robotframework/protocol.py
@@ -22,6 +22,7 @@ from ..common.parts.document_symbols import symbol_information_label
 from ..common.protocol import LanguageServerProtocol
 from .parts.codelens import RobotCodeLensProtocolPart
 from .parts.completion import RobotCompletionProtocolPart
+from .parts.debugging_utils import RobotDebuggingUtilsProtocolPart
 from .parts.diagnostics import RobotDiagnosticsProtocolPart
 from .parts.discovering import DiscoveringProtocolPart
 from .parts.document_highlight import RobotDocumentHighlightProtocolPart
@@ -87,12 +88,14 @@ class RobotLanguageServerProtocol(LanguageServerProtocol):
     robot_document_symbols = ProtocolPartDescriptor(RobotDocumentSymbolsProtocolPart)
     robot_robocop_diagnostics = ProtocolPartDescriptor(RobotRoboCopDiagnosticsProtocolPart)
     robot_formatting = ProtocolPartDescriptor(RobotFormattingProtocolPart)
-    robot_discovering = ProtocolPartDescriptor(DiscoveringProtocolPart)
     robot_semantic_tokens = ProtocolPartDescriptor(RobotSemanticTokenProtocolPart)
     robot_references = ProtocolPartDescriptor(RobotReferencesProtocolPart)
     robot_document_highlight = ProtocolPartDescriptor(RobotDocumentHighlightProtocolPart)
     robot_codelens = ProtocolPartDescriptor(RobotCodeLensProtocolPart)
     robot_workspace = ProtocolPartDescriptor(RobotWorkspaceProtocolPart)
+
+    robot_discovering = ProtocolPartDescriptor(DiscoveringProtocolPart)
+    robot_debuggin_utils = ProtocolPartDescriptor(RobotDebuggingUtilsProtocolPart)
 
     name = "RobotCode"
     version = __version__